app/microservices/books/requirements.txt
grpc-interceptor >= 0.13.0
grpcio-tools ~= 1.30
pytest ~= 5.4
pymongo~=4.3.3
python_dateutil == 2.6.0
prometheus-client